github: Add setup for running libcxx's tests on the built toolchain
Keep this disabled by default, so that it's easy to enable temporarily
for testing occasionally.

Libcxx does test the mingw configuration in their own CI (with an older
pinned release of the compiler), so this is mostly redundant, and
primarily relevant for testing that the very latest compiler (and
python setup) works for the tests.