github: Build two macOS toolchains in serial
First build the one for the native architecture, which is built
as a normal toolchain.

Then secondly, one is cross-built for the foreign architecture
(arm64). As this one can't be executed on the host it is built,
one can't use it for building runtimes like in regular native
builds. Therefore, it has to be assembled together with runtimes
built in a regular native toolchain build.

For simplicity, it is built in chain with the native macOS
build, but it could also be chained after e.g. a linux build.