build-make: Bump to GNU make 4.4
Versions 4.3 and above only come packaged as .tar.gz and .tar.lz,
and the lzip decompressor is less commonly readily available, thus
switch back to the .tar.gz format.

Building GNU make 4.4 requires building with
-Wno-error=implicit-function-declaration. Make checks for the
stpcpy function and provides a definition if it isn't available
(and uses it in a couple places within the code base). GNU make
also uses gnulib for a few functions, and out of them,
lib/concat-filename.c uses stpcpy, but doesn't include GNU make's
own internal header which provides the fallback declaration
of the function.

Therefore, either Make should use gnulib's replacement of stpcpy
(which is provided in a string.h header overlaid on top of the
platform's headers), which would be available in other gnulib
functions, or the gnulib concat-filename.c function should be
fixed not to rely on possibly-missing functions like stpcpy,
to make the gnulib functions more independent.